Initial check-in of functionality to write to Firebase

+++ Server/src/Domain/Constants/Convert.cs
@@ -4,9 +4,9 @@ namespace ScotlandsMountains.Domain.Calculations
 {
     public static class Convert
     {
-        private const decimal FeetPerMetre = 3.28084m;
+        private const double FeetPerMetre = 3.28084;
 
-        public static decimal ToFeetFrom(decimal metres, int? decimalPlaces = 0)
+        public static double ToFeetFrom(double metres, int? decimalPlaces = 0)
         {
             return Math.Round(metres*FeetPerMetre, decimalPlaces ?? 0, MidpointRounding.AwayFromZero);
         }
