+++ Server/src/Domain/ValueTypes/Height.cs
@@ -4,8 +4,8 @@ namespace ScotlandsMountains.Domain.ValueTypes
 {
     public class Height
     {
-        public decimal Metres { get; set; }
+        public double Metres { get; set; }
 
-        public decimal Feet => Convert.ToFeetFrom(Metres);
+        public double Feet => Convert.ToFeetFrom(Metres);
     }
 }
